Modify EventRegistrationList module
The custom SQL Stored Procedure (SP) used in the module was not
returning
Promotions unless they were marked with a Primary or Secondary Topic
Area in the Event's Public Calendar settings. This modification no
longer uses
the SP in favor of using a method from the Arena class,
"PromotionRequestData."

The method, "GetCurrentPromotionWebRequests_DT" takes the following
arguments:
- TopicAreaIDs (a comma-delimited string of Topic Area IDs)
- areaFilter ("primary", "secondary", "both")
- campusID (integer)
- maxItems (integer: max number of items returned)
- eventsOnly (boolean)
- documentTypeId (integer: the document type to display for a Promotion)
- organizationId (integer)

The method, "GetCurrentPromotionWebRequests_DT" returns the following
fields for each Promotion:
- promotion_request_id
- priority
- title
- web_summary
- summary_image_guid
- summary_image_path

This modification also begins using the method "GetPromotionRequestByID"
in the ASP Repeater's ItemDataBound event to get further information for

each Promotion, based on the "promotion_request_id" field, although this

functionality is not fully implemented in this commit.