Modify EventRegistrationList module
Removed all JavaScript that was dependent upon third-party plugins:
BlockUI and TableSorter.

Set up JQuery to create CSS classes for alternating table rows so that
they can be styled if desired. Changed the pop-up from using the BlockUI
plugin to using the built-in JQuery "dialog" function and altered the
HTML of the dialog structure accordingly.