Make jQuery versions compatible
There were compatibility problems with jQuery 1.8.0 and jQuery UI 1.8.5
in how the dialog widget behaved, so the UI version was changed to
1.8.24.

The CSS file, "jquery-ui_revisedForJFBC.css" had incorrect URL paths in
it.
The link to import it was removed from JohnsonFerryMain.css.

The JohnsonFerryMain_ArenaNav template was altered to use jQuery UI
1.8.24
and the "jquery-ui-.custom.css" file.

The EventRegistrationList module's JavaScript for the popup dialog was
fixed.