Modify EventRegistrationList module
The Max Items setting was not acting correctly. The code first pulls the
entire
list of items and then filters them by type. The Max Items setting was
being
applied to the initial list, rather than the filtered results, which was
causing the
number of rows in the Repeater to be incorrect in some cases.

The revision hard-codes a very high maximum for the initial data pull,
and
then applies the Max Items setting to the filtered results.