Make Event List use single jQuery UI
The EventRegistrationList module's dialog widget was not working when it

was placed on a page that also contained a module with the Arena
DataGrid
control in it. This was because the DataGrid control had an old version
of the jQuery UI hard-coded into it.

This solution involved placing the dialog code within a getScript()
function block. The getScript() function takes a specific script URL as
an argument, and uses that script when running its function.